Move remote_metric_runner into aidrin; add endpoint version check
The Globus Compute runner lived in web/globus.py, so its __module__ was
"web.globus". Globus serialises functions by reference, so the remote
endpoint worker tried to `import web` during deserialisation and failed
with `ModuleNotFoundError: No module named 'web'` — the endpoint has
aidrin installed but not the web app.

Move remote_metric_runner (already self-contained) into
aidrin.compute.remote so the reference resolves against a package the
endpoint actually has. Re-export from web.globus for backward
compatibility.

Add remote_env_probe() plus check_endpoint_compatibility() and a
/globus/check-endpoint route: on connect, run the probe on the endpoint
and compare aidrin/Python versions against the web server. An aidrin
major.minor mismatch is treated as incompatible; a Python minor
difference is a warning. The probe doubles as an import/reachability
check, turning cryptic dill failures into a clear message at connect
time.
